+++ app/src/main/java/com/example/iplmarket_fe/Chatting.java
@@ -1,4 +1,4 @@
-package com.example.iplmarket_fe.home;
+package com.example.iplmarket_fe;
 
 import android.os.Bundle;
