+++ app/src/main/java/com/example/iplmarket_fe/IdCheckResponse.java
@@ -1,0 +1,14 @@
+package com.example.iplmarket_fe;
+
+public class IdCheckResponse {
+    private boolean success;
+    private String message;
+
+    public boolean isSuccess() {
+        return success;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
