+++ app/src/main/java/com/example/iplmarket_fe/LoginData.java
@@ -1,0 +1,16 @@
+package com.example.iplmarket_fe;
+
+import com.google.gson.annotations.SerializedName;
+
+public class LoginData {
+    @SerializedName("userEmail")
+    String userEmail;
+
+    @SerializedName("userPwd")
+    String userPwd;
+
+    public LoginData(String userEmail, String userPwd) {
+        this.userEmail = userEmail;
+        this.userPwd = userPwd;
+    }
+}
