+++ app/src/main/java/com/example/iplmarket_fe/PageMain.java
@@ -1,4 +1,4 @@
-package com.example.iplmarket_fe.setting;
+package com.example.iplmarket_fe;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
