+++ app/src/main/java/com/example/iplmarket_fe/Product.java
@@ -1,4 +1,4 @@
-package com.example.iplmarket_fe.home;
+package com.example.iplmarket_fe;
 
 import android.media.Image;
