+++ app/src/main/java/com/example/iplmarket_fe/RegisterData.java
@@ -1,0 +1,27 @@
+package com.example.iplmarket_fe;
+
+import com.google.gson.annotations.SerializedName;
+
+public class RegisterData {
+    @SerializedName("userName")
+    private String userName;
+
+    @SerializedName("userPwd")
+    private String userPwd;
+
+    @SerializedName("id")
+    private String id;
+
+    public RegisterData(String userName, String userPwd) {
+        this.userName = userName;
+        this.userPwd = userPwd;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+}
