+++ app/src/main/java/com/example/iplmarket_fe/RegisterResponse.java
@@ -1,0 +1,25 @@
+package com.example.iplmarket_fe;
+
+import com.google.gson.annotations.SerializedName;
+
+public class RegisterResponse {
+    @SerializedName("code")
+    private int code;
+
+    @SerializedName("message")
+    private String message;
+    @SerializedName("success")
+    private boolean success;
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public boolean isSuccess() {
+        return success;
+    }
+}
