+++ app/src/main/java/com/example/iplmarket_fe/RetrofitClient.java
@@ -1,0 +1,16 @@
+package com.example.iplmarket_fe;
+
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class RetrofitClient {
+    private final static String BASE_URL = "http://192.168.35.91:8080/";
+    private static Retrofit retrofit = new Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build();
+
+    public static Retrofit getClient() {
+        return retrofit;
+    }
+}
