+++ app/src/main/java/com/example/iplmarket_fe/Wrote.java
@@ -1,4 +1,4 @@
-package com.example.iplmarket_fe.setting;
+package com.example.iplmarket_fe;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,22 +14,19 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.iplmarket_fe.R;
-import com.example.iplmarket_fe.home.Detail;
-
 import java.util.ArrayList;
 import java.util.List;
 
-public class Wishlist extends AppCompatActivity {
+public class Wrote extends AppCompatActivity {
 
-    private List<Product> productList;
+    private List<ProductItem> productList;
     private RecyclerView recyclerView;
     private ProductAdapter productAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_wishlist);
+        setContentView(R.layout.activity_wrote);
 
         recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
@@ -38,31 +35,31 @@ public class Wishlist extends AppCompatActivity {
         recyclerView.setAdapter(productAdapter);
 
         // 예시 데이터 추가
-        productList.add(new Product(R.drawable.product1, "Product 1", "10,000₩", false));
-        productList.add(new Product(R.drawable.product2, "Product 2", "20,000₩", false));
-        productList.add(new Product(R.drawable.product3, "상품 3", "30,000₩", false));
-        productList.add(new Product(R.drawable.product4, "상품 4", "40,000₩", false));
-        productList.add(new Product(R.drawable.product5, "상품 5", "50,000₩", false));
-        productList.add(new Product(R.drawable.product6, "상품 6", "60,000₩", false));
-        productList.add(new Product(R.drawable.product7, "상품 7", "70,000₩", false));
-        productList.add(new Product(R.drawable.product8, "상품 8", "80,000₩", false));
-        productList.add(new Product(R.drawable.product9, "상품 9", "90,000₩", false));
-        productList.add(new Product(R.drawable.product10, "상품 10", "100,000₩", false));
+        productList.add(new ProductItem(R.drawable.product1, "상품 1", "10,000₩", false));
+        productList.add(new ProductItem(R.drawable.product2, "상품 2", "20,000₩", false));
+        productList.add(new ProductItem(R.drawable.product3, "상품 3", "30,000₩", false));
+        productList.add(new ProductItem(R.drawable.product4, "상품 4", "40,000₩", false));
+        productList.add(new ProductItem(R.drawable.product5, "상품 5", "50,000₩", false));
+        productList.add(new ProductItem(R.drawable.product6, "상품 6", "60,000₩", false));
+        productList.add(new ProductItem(R.drawable.product7, "상품 7", "70,000₩", false));
+        productList.add(new ProductItem(R.drawable.product8, "상품 8", "80,000₩", false));
+        productList.add(new ProductItem(R.drawable.product9, "상품 9", "90,000₩", false));
+        productList.add(new ProductItem(R.drawable.product10, "상품 10", "100,000₩", false));
 
         productAdapter.notifyDataSetChanged();
     }
 
-    private class Product {
+    private class ProductItem {
         private int imageResource;
         private String name;
         private String price;
-        private boolean liked;
+        private boolean wrote;
 
-        public Product(int imageResource, String name, String price, boolean liked) {
+        public ProductItem(int imageResource, String name, String price, boolean wrote) {
             this.imageResource = imageResource;
             this.name = name;
             this.price = price;
-            this.liked = liked;
+            this.wrote = wrote;
         }
 
         public int getImageResource() {
@@ -77,62 +74,52 @@ public class Wishlist extends AppCompatActivity {
             return price;
         }
 
-        public boolean isLiked() {
-            return liked;
+        public boolean isWrote() {
+            return wrote;
         }
 
-        public void setLiked(boolean liked) {
-            this.liked = liked;
+        public void setWrote(boolean wrote) {
+            this.wrote = wrote;
         }
     }
 
     private class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ProductViewHolder> {
-        private List<Product> productList;
+        private List<ProductItem> productList;
 
-        public ProductAdapter(List<Product> productList) {
+        public ProductAdapter(List<ProductItem> productList) {
             this.productList = productList;
         }
 
         @NonNull
         @Override
         public ProductViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.product_item, parent, false);
+            View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.product_itemm, parent, false);
             return new ProductViewHolder(view);
         }
 
         @Override
         public void onBindViewHolder(@NonNull ProductViewHolder holder, int position) {
-            final Product product = productList.get(position);
+            final ProductItem product = productList.get(position);
 
             holder.productImage.setImageResource(product.getImageResource());
             holder.productName.setText(product.getName());
             holder.productPrice.setText(product.getPrice());
 
-            if (product.isLiked()) {
-                holder.likeButton.setImageResource(R.drawable.heart_after);
-            } else {
-                holder.likeButton.setImageResource(R.drawable.heart_before);
-            }
-
             holder.itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    // Navigate to ProductPage
-                    Intent intent = new Intent(Wishlist.this, Detail.class);
+                    // 상품 페이지로 이동
+                    Intent intent = new Intent(Wrote.this, Detail.class);
                     startActivity(intent);
                 }
             });
 
-            holder.likeButton.setOnClickListener(new View.OnClickListener() {
+            holder.wroteButton.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    if (product.isLiked()) {
-                        holder.likeButton.setImageResource(R.drawable.heart_before);
-                        product.setLiked(false);
-                    } else {
-                        holder.likeButton.setImageResource(R.drawable.heart_after);
-                        product.setLiked(true);
-                    }
+                    // 메인 화면으로 이동
+                    Intent intent = new Intent(Wrote.this, MainActivity.class);
+                    startActivity(intent);
                 }
             });
         }
@@ -146,14 +133,14 @@ public class Wishlist extends AppCompatActivity {
             private ImageView productImage;
             private TextView productName;
             private TextView productPrice;
-            private ImageButton likeButton;
+            private ImageButton wroteButton;
 
             public ProductViewHolder(@NonNull View itemView) {
                 super(itemView);
                 productImage = itemView.findViewById(R.id.productimage1);
                 productName = itemView.findViewById(R.id.productName);
                 productPrice = itemView.findViewById(R.id.productPrice);
-                likeButton = itemView.findViewById(R.id.likeButton);
+                wroteButton = itemView.findViewById(R.id.productwrote);
             }
         }
     }
