+++ app/src/main/res/layout/activity_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".setting.Register"
+    tools:context=".Register"
     android:orientation="vertical">
 
     <TextView
@@ -28,6 +28,8 @@
             android:id="@+id/reg_id"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:imeOptions="actionNext"
             android:hint="아이디"
             android:maxLength="10"/>
 
@@ -56,6 +58,8 @@
             android:id="@+id/reg_pw"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:imeOptions="actionGo"
             android:hint="비밀번호"
             android:inputType="textPassword"
             android:maxLength="15" />
@@ -81,6 +85,8 @@
             android:id="@+id/reg_name"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:imeOptions="actionGo"
             android:hint="이름"/>
 
     </com.google.android.material.textfield.TextInputLayout>
@@ -96,6 +102,8 @@
             android:id="@+id/reg_nick"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:imeOptions="actionGo"
             android:hint="닉네임"
             android:maxLength="12"/>
 
@@ -119,6 +127,8 @@
             android:id="@+id/reg_number"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:singleLine="true"
+            android:imeOptions="actionGo"
             android:hint="전화번호" />
 
     </com.google.android.material.textfield.TextInputLayout>
@@ -127,6 +137,8 @@
         android:id="@+id/reg_btn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:singleLine="true"
+        android:imeOptions="actionDone"
         android:text="회원가입"
         android:layout_gravity="center"
         android:layout_marginTop="40dp"
